# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CategorySummary {
  name: string;
  count: number;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  /** @nullable */
  originalPrice: number | null;
  category: string;
  image: string;
  rating: number;
  reviews: number;
}

export interface ProductPage {
  items: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderInput {
  /** @minItems 1 */
  items: OrderItemInput[];
}

export interface OrderItem {
  id: number;
  /** @nullable */
  productId: number | null;
  productName: string;
  productImage: string;
  category: string;
  quantity: number;
  priceAtTime: number;
}

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  subtotal: number;
  shipping: number;
  tax: number;
  total: number;
  createdAt: string;
  items: OrderItem[];
}

export type ListProductsParams = {
category?: string;
search?: string;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};