# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
  - name: orders
    description: Order operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /categories:
    get:
      operationId: listCategories
      tags: [products]
      summary: List product categories with counts
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySummary"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filtering, search, and pagination
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 24
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List the current user's orders, most recent first
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order from a list of product ids and quantities
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      description: Clerk session (cookie on web, bearer token on mobile)
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CategorySummary:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer
      required:
        - name
        - count
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        image:
          type: string
        rating:
          type: number
        reviews:
          type: integer
      required:
        - id
        - name
        - price
        - originalPrice
        - category
        - image
        - rating
        - reviews
    ProductPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
      required:
        - productId
        - quantity
    OrderInput:
      type: object
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - items
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        productImage:
          type: string
        category:
          type: string
        quantity:
          type: integer
        priceAtTime:
          type: number
      required:
        - id
        - productId
        - productName
        - productImage
        - category
        - quantity
        - priceAtTime
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
        subtotal:
          type: number
        shipping:
          type: number
        tax:
          type: number
        total:
          type: number
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - orderNumber
        - status
        - subtotal
        - shipping
        - tax
        - total
        - createdAt
        - items