# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: number;
  /** @nullable */
  productId: number | null;
  productName: string;
  productImage: string;
  category: string;
  quantity: number;
  priceAtTime: number;
}