# lib/api-zod/src/generated/types/orderItemInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}