# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  price: number;
  /** @nullable */
  originalPrice: number | null;
  category: string;
  image: string;
  rating: number;
  reviews: number;
}