# lib/api-zod/src/generated/types/productPage.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface ProductPage {
  items: Product[];
  total: number;
  page: number;
  limit: number;
}